Add TTCN3ActionRules and TTCN3Compiler
TTCN3Compiler, will contain the defintion of the specific nodes for the
AST (inheriting from List), and it will actually perform the actions
depending on the rules defined for the compiler.
The compiler will annotate after the parsing (applyRules) the action
that should be taken. In this manner, it is possible to extend the
liposuction script to do different types of refactorings and
simplifications.

TTCN3ActionRules is a type that supports the concept of the rule, a
function or predicate, that evaulates to true or false, and an action,
that should be understandable by the TTCN3Compiler.