Partial support for the select/case ttcn3 statement
I'm thinking now that it may be good to try pyparse. I kind of like
lepl, but I see some limitations on the error handling (or perhaps I
can't see the idea that its creator had in mind).
I'm also having a hard time ignoring the new lines between different
sentences. I have tried a lot of ideas but nothing.
So in the end, what I'm able to parse is just:
 'select(expression) { case (inlineTemplate) example_block }'

... I almost forgot, in this commit, I changed again the order of the
grammar productions.